pyproject.toml
[project]
name = "litellm-key-portal"
version = "0.1.0"
description = "FastAPI self-service portal for issuing LiteLLM API keys."
requires-python = ">=3.12"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.6",
  "httpx==0.27.2",
  "asyncpg==0.29.0",
  "slowapi==0.1.9",
  "python-dotenv==1.0.1",
  "python-multipart==0.0.12",
]

[dependency-groups]
dev = [
  "pytest==8.3.2",
  "pytest-asyncio==0.24.0",
  "anyio==4.4.0",
]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
